AcEmulationControl end
