+++ src/main/java/com/powsybl/openloadflow/network/LfHvdc.java
@@ -7,12 +7,63 @@
  */
 package com.powsybl.openloadflow.network;
 
+import com.powsybl.iidm.network.TwoSides;
 import com.powsybl.openloadflow.util.Evaluable;
+import com.powsybl.openloadflow.util.PerUnit;
 
 /**
  * @author Anne Tilloy {@literal <anne.tilloy at rte-france.com>}
  */
 public interface LfHvdc extends LfElement {
+
+    class AcEmulationControl {
+        private final double droop;
+        private final double p0;
+        private final double pMaxFromCS1toCS2;
+        private final double pMaxFromCS2toCS1;
+        private boolean activated = true;
+        private TwoSides feedingSide;
+
+        public AcEmulationControl(double droop, double p0, double pMaxFromCS1toCS2, double pMaxFromCS2toCS1) {
+            this.droop = droop;
+            this.p0 = p0;
+            this.pMaxFromCS1toCS2 = pMaxFromCS1toCS2;
+            this.pMaxFromCS2toCS1 = pMaxFromCS2toCS1;
+            this.feedingSide = (p0 >= 0) ? TwoSides.ONE : TwoSides.TWO;
+        }
+
+        public double getDroop() {
+            return droop / PerUnit.SB;
+        }
+
+        public double getP0() {
+            return p0 / PerUnit.SB;
+        }
+
+        public double getPMaxFromCS1toCS2() {
+            return pMaxFromCS1toCS2 / PerUnit.SB;
+        }
+
+        public double getPMaxFromCS2toCS1() {
+            return pMaxFromCS2toCS1 / PerUnit.SB;
+        }
+
+        public boolean isActivated() {
+            return activated;
+        }
+
+        public TwoSides getFeedingSide() {
+            return feedingSide;
+        }
+
+        void setActivated(boolean activated) {
+            this.activated = activated;
+        }
+
+        void setFeedingSide(TwoSides side) {
+            feedingSide = side;
+        }
+    }
 
     LfBus getBus1();
 
@@ -39,6 +90,8 @@ public interface LfHvdc extends LfElement {
     void setConverterStation1(LfVscConverterStation converterStation1);
 
     void setConverterStation2(LfVscConverterStation converterStation2);
+
+    AcEmulationControl getAcEmulationControl();
 
     void updateState();
 }
