+++ src/test/java/com/powsybl/openloadflow/EquationsTest.java
@@ -17,6 +17,7 @@ import com.powsybl.openloadflow.dc.equations.DcVariableType;
 import com.powsybl.openloadflow.equations.*;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.network.impl.LfVscConverterStationImpl;
+import com.powsybl.openloadflow.util.PerUnit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -63,6 +64,7 @@ class EquationsTest {
     private static final double LOSS_FACTOR_1 = 0.01100000023841858;
     private static final double LOSS_FACTOR_2 = 0.02400453453002384;
     private static final double G_SHUNT = 0.0000372472384299244;
+    private static final LfHvdc.AcEmulationControl AC_EMULATION_CONTROL = new LfHvdc.AcEmulationControl(DROOP * PerUnit.SB, P_0 * PerUnit.SB, Double.MAX_VALUE, Double.MAX_VALUE);
 
     private static <V extends Enum<V> & Quantity, E extends Enum<E> & Quantity> double[] eval(EquationTerm<V, E> term, List<Variable<V>> variables, StateVector sv) {
         term.setStateVector(sv);
@@ -255,10 +257,7 @@ class EquationsTest {
         var hvdc = Mockito.mock(LfHvdc.class, new RuntimeExceptionAnswer());
         Mockito.doReturn(0).when(hvdc).getNum();
         Mockito.doReturn(false).when(hvdc).isDisabled();
-        Mockito.doReturn(DROOP).when(hvdc).getDroop();
-        Mockito.doReturn(P_0).when(hvdc).getP0();
-        Mockito.doReturn(Double.MAX_VALUE).when(hvdc).getPMaxFromCS1toCS2();
-        Mockito.doReturn(Double.MAX_VALUE).when(hvdc).getPMaxFromCS2toCS1();
+        Mockito.doReturn(AC_EMULATION_CONTROL).when(hvdc).getAcEmulationControl();
         LfVscConverterStationImpl station1 = Mockito.mock(LfVscConverterStationImpl.class, new RuntimeExceptionAnswer());
         LfVscConverterStationImpl station2 = Mockito.mock(LfVscConverterStationImpl.class, new RuntimeExceptionAnswer());
         Mockito.doReturn(station1).when(hvdc).getConverterStation1();
